rounding to make display nicer

+++ AmpTools/GPUManager/GPUManager.cc
@@ -196,7 +196,7 @@ GPUManager::init( const AmpVecs& a, bool use4Vectors )
 
   totalMemory /= (1024*1024);
   
-  cout << "Attempting to allocate " << totalMemory << " MB of global GPU memory." << endl;
+  cout << "Attempting to allocate " << (int)totalMemory << " MB of global GPU memory." << endl;
   
   // device memory needed for intensity or integral calculation and sum
   gpuErrChk( cudaMalloc( (void**)&m_pfDevVVStar  , m_iVArrSize       ) ) ;
